Reset URL params on hide
This moves default state to a const and pushes the columns
that were found upon mount to that default state const. This
allows us to show/hide the chart, wiping the URL params so
that it doesn't show back up on page reload.

RE: https://github.com/simonw/datasette-vega/issues/36